#31 refactored project such that both pixelators use the same LCM and GCD calculus, and that if the DOCTEST_INCLUDE_DIR is not set (or BUILD_TESTING is switched off), test projects are not being configured and thus, doctest is not needed.
