+++ utilities/CMakeLists.txt
@@ -6,11 +6,18 @@ add_library( utilities
   ArgumentParser.h
   logging.cpp
   ArgumentParser.cpp
+  calculus.h
+  calculus.cpp
 )
-target_include_directories( utilities PUBLIC "${DOCTEST_INCLUDE_DIR}" )
-target_compile_definitions( utilities PRIVATE -DDOCTEST_CONFIG_DISABLE)
 
-add_executable( argparser_test ArgumentParser.cpp )
-target_include_directories( argparser_test PUBLIC ${CMAKE_SOURCE_DIR}/utilities )
-target_link_libraries( argparser_test PUBLIC utilities )
-target_compile_definitions( argparser_test PRIVATE -DDOCTEST_CONFIG_IMPLEMENT_WITH_MAIN )
+if(${BUILD_TESTING})
+  add_executable( test_argparser ArgumentParser.cpp )
+  target_include_directories( test_argparser PUBLIC ${CMAKE_SOURCE_DIR}/utilities "${DOCTEST_INCLUDE_DIR}" )
+  target_link_libraries( test_argparser PUBLIC utilities )
+  target_compile_definitions( test_argparser PRIVATE -DDOCTEST_CONFIG_IMPLEMENT_WITH_MAIN )
+  
+  add_executable( test_calculus calculus.cpp )
+  target_include_directories( test_calculus PUBLIC ${CMAKE_SOURCE_DIR}/utilities "${DOCTEST_INCLUDE_DIR}" )
+  target_link_libraries( test_calculus PUBLIC utilities )
+  target_compile_definitions( test_calculus PRIVATE -DDOCTEST_CONFIG_IMPLEMENT_WITH_MAIN )
+endif()
